pyproject.toml
[tool.poetry]
name = "obsidian-kb"
version = "0.1.0"
description = "Will use Obsidina as Knowledge base"
authors = ["Prabha Arivalagan"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
openai = "^1.23.2"
langchain-core = "^0.1.45"
langchain-community = "^0.0.34"
unstructured = "^0.13.3"
markdown = "^3.6"
langchain-openai = "^0.1.3"
python-dotenv = "^1.0.1"
pinecone-client = "^3.2.2"
streamlit = "^1.33.0"
langchain-pinecone = "^0.1.0"
langchain = "^0.1.16"
cohere = "^5.3.2"
langchain-cohere = "^0.1.3"
langsmith = "^0.1.49"
ragas = "^0.1.7"
numpy = "^1.26.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
